ml-service/requirements.txt
Flask==3.0.0
pandas
prophet
pymongo
python-dotenv
flask-cors